Update README.md
The well-known reinforcement learning problem Mountain Car is frequently used as a standard to evaluate how well algorithms for reinforcement learning perform. In order to get to the goal at the top of the hill, an underpowered car must drive up a steep hill in this problem. The car cannot just drive straight up the hill because it is subject to the laws of physics. Rather, it needs to accelerate up and down the hill to increase its speed.

The position and velocity of the car serve as continuous variables that represent the state of the Mountain Car environment. The agent's objective is to become proficient at controlling the car's acceleration in order to climb the hill and get to the destination as soon as possible while utilizing the least quantity of energy. The objective is to reduce the amount of time steps needed to accomplish the goal because the agent receives a negative reward for each time step that is needed.

Because the reward signal is sparse and delayed, and the state space is continuous, Mountain Car presents a difficult problem for reinforcement learning. This implies that even though the reward signal is not available right away, the agent still needs to learn how to plan ahead and effectively explore the state space in order to accomplish the goal. This project implements and experiments with Q-learning on the Mountain Car benchmark. It explores the impact of different parameters on the Q-learning algorithm's performance and reports the findings